Update PID_motor_Rotation.py
added int to theta -= (theta/360) * 360

+++ PID_motor_Rotation.py
@@ -214,7 +214,7 @@ def Odometry():
     TotalTravel += travel
     thetaRad += (leftTravel - rightTravel)/wheelbase
     theta = thetaRad*(180/math.pi) #convert to heading in degrees;
-    theta -= (theta/360) * 360 # clip theta to plus or minus 360 degrees
+    theta -= int(theta/360) * 360 # clip theta to plus or minus 360 degrees
     botX += travel * math.sin(thetaRad);
     botY += travel * math.cos(thetaRad);
     print(int(botX), int(botY), int(theta))
